requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
scikit-learn
faiss-cpu
python-dotenv
jinja2
